+++ app/lib/ui/components/checkbox/checkbox_item_demo_screen.dart
@@ -20,6 +20,7 @@ import 'package:ouds_flutter_demo/ui/components/control_item/control_item_custom
 import 'package:ouds_flutter_demo/ui/components/control_item/control_item_customization_utils.dart';
 import 'package:ouds_flutter_demo/ui/components/control_item/control_item_enum.dart';
 import 'package:ouds_flutter_demo/ui/theme/theme_controller.dart';
+import 'package:ouds_flutter_demo/ui/utilities/app_assets.dart';
 import 'package:ouds_flutter_demo/ui/utilities/code.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_section.dart';
 import 'package:ouds_flutter_demo/ui/utilities/customizable/customizable_switch.dart';
@@ -136,7 +137,7 @@ class _CheckboxItemDemoState extends State<_CheckboxItemDemo> {
               helperTitle: ControlItemCustomizationUtils.getHelperLabelText(customizationState!),
               reversed: customizationState!.hasReversed ? true : false,
               readOnly: customizationState!.hasReadOnly ? true : false,
-              icon: customizationState!.hasIcon ? 'assets/ic_heart.svg' : null,
+              icon: customizationState!.hasIcon ? AppAssets.icons.icHeart : null,
               isError: customizationState!.hasError ? true : false,
               divider: customizationState!.hasDivider ? true : false,
               tristate: widget.indeterminate,
@@ -155,7 +156,7 @@ class _CheckboxItemDemoState extends State<_CheckboxItemDemo> {
             helperTitle: ControlItemCustomizationUtils.getHelperLabelText(customizationState!),
             reversed: customizationState!.hasReversed ? true : false,
             readOnly: customizationState!.hasReadOnly ? true : false,
-            icon: customizationState!.hasIcon ? 'assets/ic_heart.svg' : null,
+            icon: customizationState!.hasIcon ? AppAssets.icons.icHeart : null,
             isError: customizationState!.hasError ? true : false,
             divider: customizationState!.hasDivider ? true : false,
             tristate: widget.indeterminate,