+++ ouds_core/lib/components/divider/ouds_divider.dart
@@ -27,13 +27,92 @@ import 'package:ouds_theme_contract/ouds_theme.dart';
 /// - [color] The color of the divider, chosen from the [OudsDividerColor] enum.
 ///   Default value set to [OudsDividerColor.defaultColor].
 /// - [length] The length of the divider (width for horizontal, height for vertical).
-///   Defaults to [double.infinity].
+///   Defaults to [double.infinity] for horizontal and 0 for vertical.
 /// - [thickness] The thickness of the divider (height for horizontal, width for vertical).
-///   Defaults to `1.0`.
 /// - [margin] Optional margin around the divider, for spacing from surrounding content.
 ///
+///
+/// # You can use [OudsDivider.horizontal] like this :
+///
+/// ## Divider horizontal with default color :
+///
+/// Separates stacked vertical sections. Example use case: Between main content and a footer.
+///
+///
+/// ```dart
+///OudsDivider.horizontal(
+///       color: OudsDividerColor.defaultColor
+///       )
+/// ```
+///
+///
+///
+/// ## Divider vertical with default color :
+///
+/// Separates horizontally aligned elements. Example use case: Between two columns in a layout.
+///
+/// ```dart
+///OudsDivider.vertical(
+///       color: OudsDividerColor.defaultColor
+///       )
+/// ```
+///
+
+class OudsDivider extends StatelessWidget {
+  final Axis orientation;
+  final OudsDividerColor color;
+  final double length;
+  final double? thickness;
+  final EdgeInsetsGeometry? margin;
+
+  /// Creates an horizontal divider.
+  const OudsDivider.horizontal({
+    super.key,
+    this.color = OudsDividerColor.defaultColor,
+    this.length = double.infinity,
+    this.thickness,
+    this.margin,
+  }) : orientation = Axis.horizontal;
+
+  /// Creates a vertical divider.
+  const OudsDivider.vertical({
+    super.key,
+    this.color = OudsDividerColor.defaultColor,
+    this.length = 0,
+    this.thickness,
+    this.margin,
+  }) : orientation = Axis.vertical;
+
+  @override
+  Widget build(BuildContext context) {
+    final actualThickness = thickness ?? OudsTheme.of(context).componentsTokens(context).divider.borderWidth;
+
+    final divider = Container(
+      color: color.getColor(context),
+      width: orientation == Axis.horizontal ? length : actualThickness,
+      height: orientation == Axis.horizontal ? actualThickness : 50,
+      margin: margin,
+    );
+
+    return Padding(padding: EdgeInsetsDirectional.all(OudsTheme.of(context).spaceScheme(context).fixedMedium), child: divider);
+  }
+}
 
 /// Represents the available colors for [OudsDivider].
+///
+/// The authorized colors are :
+///
+/// - color-border-default,
+/// - color-border-muted,
+/// - color-border-emphasized,
+/// - color-border-brand-primary (secondary/tertiary),
+/// - color-border-on-brand-primary (secondary/tertiary),
+/// - color-always-black,
+/// - color-always-white,
+/// - color-always-on-black,
+/// - color-always-on-white
+///
+///
 enum OudsDividerColor {
   defaultColor,
   muted,
@@ -68,45 +147,5 @@ enum OudsDividerColor {
       default:
         return theme.colorScheme(context).borderDefault;
     }
-  }
-}
-
-class OudsDivider extends StatelessWidget {
-  final Axis orientation;
-  final OudsDividerColor color;
-  final double length;
-  final double? thickness;
-  final EdgeInsetsGeometry? margin;
-
-  /// Creates a horizontal divider.
-  const OudsDivider.horizontal({
-    super.key,
-    this.color = OudsDividerColor.defaultColor,
-    this.length = double.infinity,
-    this.thickness,
-    this.margin,
-  }) : orientation = Axis.horizontal;
-
-  /// Creates a vertical divider.
-  const OudsDivider.vertical({
-    super.key,
-    this.color = OudsDividerColor.defaultColor,
-    this.length = 0,
-    this.thickness,
-    this.margin,
-  }) : orientation = Axis.vertical;
-
-  @override
-  Widget build(BuildContext context) {
-    final actualThickness = thickness ?? OudsTheme.of(context).componentsTokens(context).divider.borderWidth;
-
-    final divider = Container(
-      color: color.getColor(context),
-      width: orientation == Axis.horizontal ? length : actualThickness,
-      height: orientation == Axis.horizontal ? actualThickness : 50,
-      margin: margin,
-    );
-
-    return Padding(padding: EdgeInsetsDirectional.all(OudsTheme.of(context).spaceScheme(context).fixedMedium), child: divider);
   }
 }