+++ ouds_theme_orange/lib/orange_theme.dart
@@ -16,6 +16,7 @@ import 'package:ouds_theme_contract/ouds_tokens_provider.dart';
 import 'package:ouds_theme_contract/theme/scheme/color/ouds_color_scheme.dart';
 import 'package:ouds_theme_contract/theme/scheme/responsive/ouds_grid_scheme.dart';
 import 'package:ouds_theme_contract/theme/scheme/responsive/ouds_space_scheme.dart';
+import 'package:ouds_theme_contract/theme/scheme/typography/ouds_typography.dart';
 import 'package:ouds_theme_contract/theme/tokens/components/ouds_components_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_border_semantic_tokens.dart';
 import 'package:ouds_theme_contract/theme/tokens/semantic/ouds_color_semantic_tokens.dart';
@@ -112,6 +113,9 @@ class OrangeTheme implements OudsThemeContract {
   OudsGridScheme gridScheme(BuildContext context) {
     return OudsGridScheme.fromContext(context: context, gridTokens: gridSemanticTokens);
   }
+
+  @override
+  OudsTypography get typographyTokens => OudsTypography();
 
   @override
   OudsProvidersTokens providersTokens(BuildContext context) {